pyproject.toml
[project]
name = "diffusers-tuner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "diffusers>=0.37.0",
    "hydra-core>=1.3.2",
    "prodigyopt>=1.1.2",
    "safetensors>=0.7.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
]
